requirements.txt
pytube3==9.6.4
beautifulsoup4==4.9.1
selenium==3.141.0
moviepy==1.0.3
pandas==1.1.0